move test extract fixtures to conftest
